add abstract class and make class for save and read
