+++ data/board.txt
@@ -1,0 +1,4 @@
+1 2 3 4
+5 6 7 8
+9 10 0 11
+13 14 15 12
