+++ src/a_star.py
@@ -1,0 +1,5 @@
+from algorythms import Algorithm
+
+
+class Astar(Algorithm):
+    pass
