--- src/direction.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class Direction(Enum):
-    LEFT = 'L'
-    RIGHT = 'R'
-    UP = 'U'
-    DOWN = 'D'
