--- src/fifteen_puzzle.py
@@ -1,3 +1,0 @@
-class FifteenPuzzle:
-    def __init__(self, board: list):
-        self.board = board
