+++ src/main.py
@@ -1,19 +1,11 @@
-from algorythms import Algorithm
+from bfs import Bfs
+from reader import Reader
 import time
 
-board = [[1, 2, 3, 4],
-         [5, 6, 7, 8],
-         [9, 10, 0, 11],
-         [13, 14, 15, 12]]
-
-board_2 = [[0, 1, 2, 7],
-           [8, 9, 12, 10],
-           [13, 3, 6, 4],
-           [15, 14, 11, 5]]
-
-bfs = Algorithm(board_2)
+board = Reader.read()
+bfs = Bfs(board)
 begin = time.time()
-print(bfs.bfs())
+print(bfs.simulation())
 print(f"TURNS: {bfs.move_counter}")
 end = time.time()
 print(f"Time: {end - begin}")
